Move crypto utilities into dedicated folder
