+++ iam-login-service/src/main/java/it/infn/mw/iam/util/crypto/IamPbkdf2Util.java
@@ -13,13 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package it.infn.mw.iam.util;
+package it.infn.mw.iam.util.crypto;
 
-import it.infn.mw.iam.core.client.IamHmacPasswordEncoder;
+import org.springframework.security.crypto.password.Pbkdf2PasswordEncoder;
 
-public class IamHmacUtil {
-
-  private static final String MASTER_KEY = "my-super-secret-master-key-provided-by-configuration";
+public class IamPbkdf2Util {
 
   public static void main(String[] args) {
 
@@ -28,7 +26,7 @@ public class IamHmacUtil {
       System.exit(1);
     }
 
-    IamHmacPasswordEncoder encoder = new IamHmacPasswordEncoder(MASTER_KEY);
+    Pbkdf2PasswordEncoder encoder = new Pbkdf2PasswordEncoder();
 
     String encodedPwd = encoder.encode(args[0]);
     System.out.printf("Encoded secret: %s%n", encodedPwd);
@@ -38,7 +36,7 @@ public class IamHmacUtil {
     double elapsed = (System.nanoTime() - start) / 1_000_000.0;
 
     if (pwdMatched) {
-      System.out.printf("Time to match secret with HMAC-SHA256 algorithm is %.3f ms%n", elapsed);
+      System.out.printf("Time to match secret with Pbkdf2 library is %.3f ms", elapsed);
     }
   }
